- fixed 64 bit count printf conflict - gcc seems to like this version of strstr5BitLHAt a bit more

+++ strstrProfiling/strstrBothLen.c
@@ -51,5 +51,5 @@ _noinline_ void Run_strstr_know_both_lengths_test()
 	}
 	double runtimeSec = EndTimer();
 	printf(" ... Done\n");
-	printf("Searches made %d. Found the string %d times. Seconds : %f\n\n", (int)searchesMade, (int)foundCount, (float)runtimeSec);
+	printf("Searches made %zd. Found the string %d times. Seconds : %f\n\n", searchesMade, (int)foundCount, (float)runtimeSec);
 }
